refactor: moved suite nodes to subpackage
