+++ compiler/src/main/kotlin/io/github/subjekt/nodes/suite/Parameter.kt
@@ -1,4 +1,4 @@
-package io.github.subjekt.nodes
+package io.github.subjekt.nodes.suite
 
 data class Parameter(
   val name: String,
