+++ compiler/src/main/kotlin/io/github/subjekt/nodes/suite/Suite.kt
@@ -1,4 +1,4 @@
-package io.github.subjekt.nodes
+package io.github.subjekt.nodes.suite
 
 import io.github.subjekt.yaml.Configuration
